Fixed code duplication in Login.js and SignUp.js
