+++ webapp/package.json
@@ -43,10 +43,13 @@
     ]
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.26.9",
+    "@babel/preset-react": "^7.26.3",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.2.0",
     "@testing-library/user-event": "^14.6.1",
     "axios-mock-adapter": "^2.1.0",
+    "babel-jest": "^29.7.0",
     "expect-puppeteer": "^9.0.2",
     "jest": "^29.7.0",
     "jest-cucumber": "^3.0.1",
