+++ webapp/src/components/LogInSignUp/SignUp.js
@@ -1,38 +1,20 @@
 // src/components/AddUser.js
 import React, { useState } from 'react';
-import axios from 'axios';
-import { Typography, TextField, Button, Snackbar, Box, FormControl, Divider } from '@mui/material';
+import { Typography, TextField, Button, Box, FormControl, Divider } from '@mui/material';
 import { SitemarkIcon } from '../CustomIcons';
 import { Card, LogInContainer } from '../CustomComponents';
-import { Link, useNavigate } from 'react-router-dom';
-import Cookies from 'js-cookie';
-
-const apiEndpoint = process.env.REACT_APP_API_ENDPOINT || 'http://localhost:8000';
+import { Link } from 'react-router-dom';
+import { useAuth } from '../hooks/useAuth';
 
 export const SignUp = () => {
   const [username, setUsername] = useState('');
   const [password, setPassword] = useState('');
-  const [error, setError] = useState('');
-  const [openSnackbar, setOpenSnackbar] = useState(false);
+  const { authenticateUser, error } = useAuth();
 
-  const navigate = useNavigate();
-
-  const addUser = async () => {
-    try {
-      const response = await axios.post(`${apiEndpoint}/adduser`, { username, password });
-      let oneHourAfter = new Date().getTime() + (1 * 60 * 60 * 1000)
-      Cookies.set('user', JSON.stringify({ username: response.data.username, token: response.data.token })
-        , { expires: oneHourAfter });
-      navigate('/home');
-      window.location.reload();
-    } catch (error) {
-      setError(error.response.data.error);
-    }
-  };
-
-  const handleCloseSnackbar = () => {
-    setOpenSnackbar(false);
-  };
+  const signUp = (e) => {
+    e.preventDefault();
+    authenticateUser('addUser', username, password)
+  } 
 
   return (
     <LogInContainer direction="column" justifyContent="space-between">
@@ -75,13 +57,14 @@ export const SignUp = () => {
               onChange={(e) => setPassword(e.target.value)}
             />
           </FormControl>
-          <Button variant="contained" color="primary" onClick={addUser}>
+          <Button 
+            variant="contained" 
+            color="primary" 
+            onClick={signUp}>
             Sign Up
           </Button>
-          <Snackbar open={openSnackbar} autoHideDuration={6000} onClose={handleCloseSnackbar} message="User added successfully" />
-          {error && (
-            <Snackbar open={!!error} autoHideDuration={6000} onClose={() => setError('')} message={`Error: ${error}`} />
-          )}
+          {error && <p style={{ color: 'red' }}>{error}</p>}
+
         </Box>
         <Divider>or</Divider>
         <Typography component="div" align="center" sx={{ marginTop: 2 }}>
